Added pyspike workdaround folder to the sandbox
The pyspike module seems like it will be useful to us here. However in
order to install it, we need things that will not work with brian2. So
instead of installing and compiling with cython, we just use their
original source code written in python. It is slower but should still be
okay here.